separate out addProject UI manipulation from screencontroller
