+++ src/controllers/Screen/index.js
@@ -1,4 +1,5 @@
 import { createAppController } from "../App/createAppController";
+import addProject from "./addProject";
 
 //todo: make svg imported - there is too much markup here...
 
@@ -39,57 +40,6 @@ const createScreenController = () => {
   [brandingInSidebar, inboxLinkInSidebar].forEach((elem) =>
     elem.addEventListener("click", () => handleNavLinkClicks("Inbox"))
   );
-};
-
-const addProject = () => {
-  const dialog = document.querySelector(".new-project-modal");
-  const InputAndErrorMessageElements = document.querySelector(
-    ".new-project-modal .top"
-  ).children;
-  const form = document.querySelector(".new-project-modal form");
-  const input = document.querySelector("#new-project-name");
-  const cancelBtn = document.querySelector(".new-project-modal .cancel-btn");
-
-  // event listeners and handlers
-  function handleSubmit(e) {
-    e.preventDefault();
-
-    // dealing with errors when project already exists
-    try {
-      app.addProject(input.value);
-      form.reset();
-      clearErrorMessage();
-      dialog.close();
-      displayProjectsInNav();
-    } catch {
-      [...InputAndErrorMessageElements].forEach((elem) =>
-        elem.classList.add("error")
-      );
-    }
-  }
-
-  function clearErrorMessage() {
-    [...InputAndErrorMessageElements].forEach((elem) =>
-      elem.classList.remove("error")
-    );
-  }
-
-  form.addEventListener("submit", handleSubmit);
-  input.addEventListener("input", clearErrorMessage);
-  cancelBtn.addEventListener("click", () => {
-    clearErrorMessage();
-    dialog.close();
-  });
-  document.addEventListener("keydown", (e) => {
-    if (e.key === "Escape") {
-      form.reset();
-      clearErrorMessage();
-      dialog.close();
-    }
-  });
-
-  // run
-  dialog.showModal();
 };
 
 const displayProjectsInNav = () => {
